// TruchasSurala/src/app/models/alevinos/lotes.propio.response.ts
// Generated by https://quicktype.io
export interface LotesPropiosResponse {
    code:         number;
    status:       string;
    lotesPropios: LotesPropio[];
}

export class LotesPropio {
    id:                     number;
    fecha_salida:           string;
    id_despacho:            number;
    numero_lote:            string;
    fecha_desove:           string;
    linea_genetica:         string;
    ovas_ml:                number;
    edad_tcu:               number;
    tamanio:                number;
    total_lote:             number;
    total_lote_propios:     number;
    tamanio_usado_alevinos: number;
    fecha_incubacion:       string;
    fecha_eclosion:         string;
    fecha_fin_aborcion:     string;
    fecha_primer_alimento:  string;
    temp_eclosion:          number;

}